ISPN-15330 BRPOP resp command
